custom-elements/concepts/custom-elements-type-test-002.html

<!DOCTYPE html>
<html>
<head>
<title>The custom element type is a sequence of characters that must match the NCName production and contain a minus character</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="The custom element type is a sequence of characters that must match the NCName production and contain a minus character">
<link rel="help" href="http://www.w3.org/TR/2013/WD-html-templates-20130214/#clearing-the-stack">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

function doTest(name) {
    
    var doc = newHTMLDocument();
    
    var XFoo = null;
    
    var thrown = false;    
    
    try {
        XFoo = doc.register(name);
    } catch (e) {
        thrown = true;
    }
        
    assert_true(thrown, "Exception should be thrown in case of attempt to register "
            + "a custom element with an illegal name '" + name + "'");
}

var parameters = [
    ['Invalid custom element name xfoo', 'xfoo'],
    ['Invalid custom element name x_foo', 'x_foo'],
    ['Invalid custom element name x.foo', 'x.foo'],
    ['Invalid custom element name x1foo', 'x1foo'],
    ['Invalid custom element name -xfoo', '-xfoo'],
    ['Invalid custom element name 1-foo', '1-foo'],
    ['Invalid custom element name 1x-foo', '1x-foo'],
    ['Invalid custom element name .-foo', '.-foo'],
    ['Invalid custom element name _foo', '_foo'],
    ['Invalid custom element name x-f!oo', 'x-f!oo']
];

generate_tests(doTest, parameters, 'Invalid custom element types');

</script>
</body>
</html>
